Database DAO and addTicketsScreen logic

+++ app/src/main/java/com/example/avc/database/entity/TicketEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "ticket_table")
 data class TicketEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "price") val price: Double,
     @ColumnInfo(name = "date") val date: Long
 )
